Fix for issue #12.
Due to the guice documentation (http://google.github.io/guice/api-docs/3.0/javadoc/com/google/inject/Binder.html) when the `toInstance` method is used guice is injecting all annotated fields of the instance when the Injector is created. For this reason the injection is performed twice. Using a provider is fixing this problem.